Created folder for each type of fragment
Created folder to organize fragment by each type:
- stock
- fii
- fixedincome
- currency
