+++ app/src/main/java/br/com/carteira/fragment/currency/CurrencyMainFragment.java
@@ -1,4 +1,4 @@
-package br.com.carteira.fragment;
+package br.com.carteira.fragment.currency;
 
 
 import android.os.Bundle;
@@ -7,9 +7,11 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import br.com.carteira.R;
+import br.com.carteira.fragment.BaseFragment;
 
-public class FiiMainFragment extends BaseFragment {
-    public FiiMainFragment() {
+public class CurrencyMainFragment extends BaseFragment {
+
+    public CurrencyMainFragment() {
         // Required empty public constructor
     }
 
@@ -17,13 +19,13 @@ public class FiiMainFragment extends BaseFragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         // Set fragment title
-        getActivity().setTitle(R.string.title_fii);
+        getActivity().setTitle(R.string.title_currency);
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_fii_main, container, false);
+        return inflater.inflate(R.layout.fragment_currency_main, container, false);
     }
 }
