+++ app/src/main/java/br/com/carteira/fragment/stock/BonificationFormFragment.java
@@ -1,11 +1,10 @@
-package br.com.carteira.fragment;
+package br.com.carteira.fragment.stock;
 
 
 import android.content.ContentValues;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.text.InputFilter;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -20,25 +19,21 @@ import java.util.Date;
 import br.com.carteira.R;
 import br.com.carteira.common.Constants;
 import br.com.carteira.data.PortfolioContract;
-import br.com.carteira.util.InputFilterDecimal;
-import br.com.carteira.util.InputFilterPercentage;
+import br.com.carteira.fragment.BaseFormFragment;
 
-
-public class BuyStockFormFragment extends BaseFormFragment {
-    private static final String LOG_TAG = BuyStockFormFragment.class.getSimpleName();
+public class BonificationFormFragment extends BaseFormFragment {
+    private static final String LOG_TAG = BonificationFormFragment.class.getSimpleName();
     private View mView;
 
     private EditText mInputSymbolView;
     private EditText mInputQuantityView;
-    private EditText mInputBuyPriceView;
-    private EditText mInputObjectiveView;
     private EditText mInputDateView;
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.menu_item_done:
-                if (addStock()) {
+                if (addBonification()) {
                     getActivity().finish();
                 }
                 return true;
@@ -50,13 +45,11 @@ public class BuyStockFormFragment extends BaseFormFragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        mView = inflater.inflate(R.layout.fragment_buy_stock_form, container, false);
+        mView = inflater.inflate(R.layout.fragment_bonification_form, container, false);
 
         mInputSymbolView = (EditText) mView.findViewById(R.id.inputSymbol);
         mInputQuantityView = (EditText) mView.findViewById(R.id.inputQuantity);
-        mInputBuyPriceView = (EditText) mView.findViewById(R.id.inputBuyPrice);
-        mInputObjectiveView = (EditText) mView.findViewById(R.id.inputObjective);
-        mInputDateView = (EditText) mView.findViewById(R.id.inputBuyDate);
+        mInputDateView = (EditText) mView.findViewById(R.id.inputBonificationDate);
 
         // Gets symbol received from selected CardView on intent
         Intent intent = getActivity().getIntent();
@@ -71,28 +64,23 @@ public class BuyStockFormFragment extends BaseFormFragment {
         mInputDateView.setText(simpleDateFormat.format(new Date()));
         mInputDateView.setOnClickListener(setDatePicker(mInputDateView));
 
-        // Adding input filters
-        mInputObjectiveView.setFilters(new InputFilter[]{ new InputFilterPercentage("0", "100")});
-        mInputBuyPriceView.setFilters(new InputFilter[]{ new InputFilterDecimal()});
         return mView;
     }
 
     // Validate inputted values and add the stock to the portfolio
-    private boolean addStock() {
+    private boolean addBonification() {
 
         // Validate for each inputted value
         boolean isValidSymbol = isValidStockSymbol(mInputSymbolView);
         boolean isValidQuantity = isValidInt(mInputQuantityView);
-        boolean isValidBuyPrice = isValidDouble(mInputBuyPriceView);
-        boolean isValidObjective = isValidPercent(mInputObjectiveView);
         boolean isValidDate = isValidDate(mInputDateView);
 
         // If all validations pass, try to add the stock
-        if (isValidSymbol && isValidQuantity && isValidBuyPrice && isValidObjective && isValidDate) {
+        if (isValidSymbol && isValidQuantity && isValidDate) {
             String inputSymbol = mInputSymbolView.getText().toString();
             int inputQuantity = Integer.parseInt(mInputQuantityView.getText().toString());
-            double buyPrice = Double.parseDouble(mInputBuyPriceView.getText().toString());
-            double inputObjective = Double.parseDouble(mInputObjectiveView.getText().toString());
+            double buyPrice = 0;
+            double inputObjective = -1;
             // Get and handle inserted date value
             String inputDate = mInputDateView.getText().toString();
             Long timestamp = DateToTimestamp(inputDate);
@@ -105,7 +93,7 @@ public class BuyStockFormFragment extends BaseFormFragment {
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_QUANTITY, inputQuantity);
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_PRICE, buyPrice);
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_TIMESTAMP, timestamp);
-            stockCV.put(PortfolioContract.StockTransaction.COLUMN_STATUS, Constants.Status.BUY);
+            stockCV.put(PortfolioContract.StockTransaction.COLUMN_STATUS, Constants.Status.BONIFICATION);
             // Adds to the database
             Uri insertedStockTransactionUri = mContext.getContentResolver().insert(PortfolioContract
                     .StockTransaction.URI,
@@ -116,13 +104,13 @@ public class BuyStockFormFragment extends BaseFormFragment {
                 Log.d(LOG_TAG, "Added stock transaction " + inputSymbol);
                 // Updates each stock table with new value: Income, Data, StockPortfolio, CompletePortfolio
                 updateStockIncomes(inputSymbol, timestamp);
-                boolean updateStockData = updateStockData(inputSymbol, inputQuantity, buyPrice, inputObjective, Constants.Status.BUY);
+                boolean updateStockData = updateStockData(inputSymbol, inputQuantity, buyPrice, inputObjective, Constants.Status.BONIFICATION);
                 if (updateStockData) {
-                    Toast.makeText(mContext, R.string.buy_stock_success, Toast.LENGTH_LONG).show();
+                    Toast.makeText(mContext, R.string.bonification_success, Toast.LENGTH_LONG).show();
                     return true;
                 }
             }
-            Toast.makeText(mContext, R.string.buy_stock_fail, Toast.LENGTH_LONG).show();
+            Toast.makeText(mContext, R.string.bonification_fail, Toast.LENGTH_LONG).show();
             return false;
         } else {
             // If validation fails, show validation error message
@@ -131,12 +119,6 @@ public class BuyStockFormFragment extends BaseFormFragment {
             }
             if(!isValidQuantity){
                 mInputQuantityView.setError(this.getString(R.string.wrong_quantity));
-            }
-            if(!isValidBuyPrice){
-                mInputBuyPriceView.setError(this.getString(R.string.wrong_price));
-            }
-            if(!isValidObjective){
-                mInputObjectiveView.setError(this.getString(R.string.wrong_percentual_objective));
             }
             if(!isValidDate){
                 mInputDateView.setError(this.getString(R.string.wrong_date));
