+++ app/src/main/java/br/com/carteira/fragment/stock/BuyStockFormFragment.java
@@ -1,4 +1,4 @@
-package br.com.carteira.fragment;
+package br.com.carteira.fragment.stock;
 
 
 import android.content.ContentValues;
@@ -20,20 +20,26 @@ import java.util.Date;
 import br.com.carteira.R;
 import br.com.carteira.common.Constants;
 import br.com.carteira.data.PortfolioContract;
+import br.com.carteira.fragment.BaseFormFragment;
+import br.com.carteira.util.InputFilterDecimal;
+import br.com.carteira.util.InputFilterPercentage;
 
-public class BonificationFormFragment extends BaseFormFragment {
-    private static final String LOG_TAG = BonificationFormFragment.class.getSimpleName();
+
+public class BuyStockFormFragment extends BaseFormFragment {
+    private static final String LOG_TAG = BuyStockFormFragment.class.getSimpleName();
     private View mView;
 
     private EditText mInputSymbolView;
     private EditText mInputQuantityView;
+    private EditText mInputBuyPriceView;
+    private EditText mInputObjectiveView;
     private EditText mInputDateView;
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.menu_item_done:
-                if (addBonification()) {
+                if (addStock()) {
                     getActivity().finish();
                 }
                 return true;
@@ -45,11 +51,13 @@ public class BonificationFormFragment extends BaseFormFragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        mView = inflater.inflate(R.layout.fragment_bonification_form, container, false);
+        mView = inflater.inflate(R.layout.fragment_buy_stock_form, container, false);
 
         mInputSymbolView = (EditText) mView.findViewById(R.id.inputSymbol);
         mInputQuantityView = (EditText) mView.findViewById(R.id.inputQuantity);
-        mInputDateView = (EditText) mView.findViewById(R.id.inputBonificationDate);
+        mInputBuyPriceView = (EditText) mView.findViewById(R.id.inputBuyPrice);
+        mInputObjectiveView = (EditText) mView.findViewById(R.id.inputObjective);
+        mInputDateView = (EditText) mView.findViewById(R.id.inputBuyDate);
 
         // Gets symbol received from selected CardView on intent
         Intent intent = getActivity().getIntent();
@@ -64,23 +72,28 @@ public class BonificationFormFragment extends BaseFormFragment {
         mInputDateView.setText(simpleDateFormat.format(new Date()));
         mInputDateView.setOnClickListener(setDatePicker(mInputDateView));
 
+        // Adding input filters
+        mInputObjectiveView.setFilters(new InputFilter[]{ new InputFilterPercentage("0", "100")});
+        mInputBuyPriceView.setFilters(new InputFilter[]{ new InputFilterDecimal()});
         return mView;
     }
 
     // Validate inputted values and add the stock to the portfolio
-    private boolean addBonification() {
+    private boolean addStock() {
 
         // Validate for each inputted value
         boolean isValidSymbol = isValidStockSymbol(mInputSymbolView);
         boolean isValidQuantity = isValidInt(mInputQuantityView);
+        boolean isValidBuyPrice = isValidDouble(mInputBuyPriceView);
+        boolean isValidObjective = isValidPercent(mInputObjectiveView);
         boolean isValidDate = isValidDate(mInputDateView);
 
         // If all validations pass, try to add the stock
-        if (isValidSymbol && isValidQuantity && isValidDate) {
+        if (isValidSymbol && isValidQuantity && isValidBuyPrice && isValidObjective && isValidDate) {
             String inputSymbol = mInputSymbolView.getText().toString();
             int inputQuantity = Integer.parseInt(mInputQuantityView.getText().toString());
-            double buyPrice = 0;
-            double inputObjective = -1;
+            double buyPrice = Double.parseDouble(mInputBuyPriceView.getText().toString());
+            double inputObjective = Double.parseDouble(mInputObjectiveView.getText().toString());
             // Get and handle inserted date value
             String inputDate = mInputDateView.getText().toString();
             Long timestamp = DateToTimestamp(inputDate);
@@ -93,7 +106,7 @@ public class BonificationFormFragment extends BaseFormFragment {
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_QUANTITY, inputQuantity);
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_PRICE, buyPrice);
             stockCV.put(PortfolioContract.StockTransaction.COLUMN_TIMESTAMP, timestamp);
-            stockCV.put(PortfolioContract.StockTransaction.COLUMN_STATUS, Constants.Status.BONIFICATION);
+            stockCV.put(PortfolioContract.StockTransaction.COLUMN_STATUS, Constants.Status.BUY);
             // Adds to the database
             Uri insertedStockTransactionUri = mContext.getContentResolver().insert(PortfolioContract
                     .StockTransaction.URI,
@@ -104,13 +117,13 @@ public class BonificationFormFragment extends BaseFormFragment {
                 Log.d(LOG_TAG, "Added stock transaction " + inputSymbol);
                 // Updates each stock table with new value: Income, Data, StockPortfolio, CompletePortfolio
                 updateStockIncomes(inputSymbol, timestamp);
-                boolean updateStockData = updateStockData(inputSymbol, inputQuantity, buyPrice, inputObjective, Constants.Status.BONIFICATION);
+                boolean updateStockData = updateStockData(inputSymbol, inputQuantity, buyPrice, inputObjective, Constants.Status.BUY);
                 if (updateStockData) {
-                    Toast.makeText(mContext, R.string.bonification_success, Toast.LENGTH_LONG).show();
+                    Toast.makeText(mContext, R.string.buy_stock_success, Toast.LENGTH_LONG).show();
                     return true;
                 }
             }
-            Toast.makeText(mContext, R.string.bonification_fail, Toast.LENGTH_LONG).show();
+            Toast.makeText(mContext, R.string.buy_stock_fail, Toast.LENGTH_LONG).show();
             return false;
         } else {
             // If validation fails, show validation error message
@@ -119,6 +132,12 @@ public class BonificationFormFragment extends BaseFormFragment {
             }
             if(!isValidQuantity){
                 mInputQuantityView.setError(this.getString(R.string.wrong_quantity));
+            }
+            if(!isValidBuyPrice){
+                mInputBuyPriceView.setError(this.getString(R.string.wrong_price));
+            }
+            if(!isValidObjective){
+                mInputObjectiveView.setError(this.getString(R.string.wrong_percentual_objective));
             }
             if(!isValidDate){
                 mInputDateView.setError(this.getString(R.string.wrong_date));
