+++ app/src/main/res/layout/fragment_currency_main.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="br.com.carteira.fragment.CurrencyMainFragment">
+    tools:context="br.com.carteira.fragment.currency.CurrencyMainFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
