+++ app/src/main/res/layout/fragment_fixed_income_main.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="br.com.carteira.fragment.FixedIncomeMainFragment">
+    tools:context="br.com.carteira.fragment.fixedincome.FixedIncomeMainFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
